List Atoms

List Atoms

List Atoms

Lookup Definitions

Lookup Relations

Lookup Source Asserted Identifiers

Lookup Source Parents

Lookup Source Children

Lookup Source Ancestors

Lookup Source Descendants

Lookup Source Relations

Lookup Source Subsets

List Source Subsets

Lookup Source Attributes

Semantic Network

List Content Views

Lookup Content Views

Lookup a Crosswalk

list_atoms(
  CUI,
  sabs = NULL,
  ttys = NULL,
  language = NULL,
  includeObsolete = NULL,
  includeSuppressible = NULL,
  pageNumber = NULL,
  pageSize = NULL
)

lookup_atoms(
  AUI,
  sabs = NULL,
  ttys = NULL,
  language = NULL,
  includeObsolete = NULL,
  includeSuppressible = NULL,
  pageNumber = NULL,
  pageSize = NULL
)

lookup_definitions(sabs = NULL, pageNumber = NULL, pageSize = NULL)

lookup_relations(pageNumber = NULL, pageSize = NULL)

lookup_source_asserted_identifiers(sourceId, sab)

lookup_source_parents(sourceId, sab, pageNumber = NULL, pageSize = NULL)

lookup_source_children(sourceId, sab, pageNumber = NULL, pageSize = NULL)

lookup_source_ancestors(pageNumber = NULL, pageSize = NULL)

lookup_source_descendants(pageNumber = NULL, pageSize = NULL)

lookup_source_relations(
  sourceId,
  vocabulary,
  includeRelationLabels = NULL,
  includeAdditionalRelationLabels = NULL,
  includeObsolete = NULL,
  includeSuppressible = NULL,
  pageNumber = NULL,
  pageSize = NULL
)

list_subsets(pageNumber = NULL, pageSize = NULL, language = "ENG")

list_source_subsets(
  vocabulary,
  pageNumber = NULL,
  pageSize = NULL,
  language = "ENG"
)

lookup_source_attributes(
  sourceId,
  vocabulary,
  pageNumber = NULL,
  pageSize = NULL,
  includeAttributeNames = NULL
)

lookup_semantic_network()

list_content_views(pageNumber = NULL, pageSize = NULL)

lookup_content_view_members(CUI, pageNumber = NULL, pageSize = NULL)

lookup_crosswalk(
  targetSource = NULL,
  includeObsolete = NULL,
  pageNumber = NULL,
  pageSize = NULL
)

Arguments

sabs
(optional) Comma-separated list of source vocabularies to include in your search

ttys
(optional) One or more term types

language
(optional) 3-letter abbreviation for language

includeObsolete
(optional) Determines whether to return obsolete codes.

includeSuppressible
(optional) Include content that is suppressible according to NLM Editors.

pageNumber
(optional) Whole number that specifies which page of results to fetch.

pageSize
(optional) Whole number that specifies the number of results to include per page.

includeRelationLabels
(optional) One or more relation labels

includeAdditionalRelationLabels
(optional) One or more relation attribute

includeAttributeNames
(optional) One or more attribute names

targetSource
(optional) Returns codes from the specified UMLS vocabulary

ticket
A single-use service ticket is required for each call to the API. See authentication for more information

Details

'/content/current/CUI/C0155502/atoms'path Retrieves all atoms for C0155502 and returns a JSON Object classType of Atom '/content/current/CUI/C0155502/atoms/preferred'path Retrieves the default preferred atom of C0155502 and returns a JSON Object classType of Atom '/content/current/CUI/C0155502/atoms?language=ENG'path Retrieves all English language atoms for C0155502 and returns a JSON Object classType of Atom '/content/current/CUI/C0155502/atoms?sabs=SNOMEDCT_US,ICD9CM&ttys=PT'path Retrieve SNOMEDCT_US and ICD9CM preferred terms in C0155502 and returns a JSON Object classType of Atom '/content/current/source/SNOMEDCT_US/111541001/atoms'path Retrieve atoms belonging to SNOMED CT concept 111541001 and returns a JSON Object classType of Atom '/content/current/source/SNOMEDCT_US/111541001/atoms/preferred'path Retrieve the default preferred atom belonging to SNOMED CT concept 111541001 and returns a JSON Object classType of Atom '/content/current/AUI/A8345234'path Retrieve information about AUI A8345234 and returns a JSON Object classType of Atom '/content/current/AUI/A8345234/ancestors'path Retrieve ancestors of a UMLS atom and returns a JSON Object classType of Atom '/content/current/AUI/A8345234/descendants'path Retrieve descendants of a UMLS atom and returns a JSON Object classType of Atom '/content/current/AUI/A8345234/parents'path Retrieve parents of a UMLS atom and returns a JSON Object classType of Atom '/content/current/AUI/A8345234/children'path Retrieve children of a UMLS atom and returns a JSON Object classType of Atom

'/content/current/CUI/C0155502/atoms'path Retrieves all atoms for C0155502 and returns a JSON Object classType of Atom '/content/current/CUI/C0155502/atoms/preferred'path Retrieves the default preferred atom of C0155502 and returns a JSON Object classType of Atom '/content/current/CUI/C0155502/atoms?language=ENG'path Retrieves all English language atoms for C0155502 and returns a JSON Object classType of Atom '/content/current/CUI/C0155502/atoms?sabs=SNOMEDCT_US,ICD9CM&ttys=PT'path Retrieve SNOMEDCT_US and ICD9CM preferred terms in C0155502 and returns a JSON Object classType of Atom '/content/current/source/SNOMEDCT_US/111541001/atoms'path Retrieve atoms belonging to SNOMED CT concept 111541001 and returns a JSON Object classType of Atom '/content/current/source/SNOMEDCT_US/111541001/atoms/preferred'path Retrieve the default preferred atom belonging to SNOMED CT concept 111541001 and returns a JSON Object classType of Atom '/content/current/AUI/A8345234'path Retrieve information about AUI A8345234 and returns a JSON Object classType of Atom '/content/current/AUI/A8345234/ancestors'path Retrieve ancestors of a UMLS atom and returns a JSON Object classType of Atom '/content/current/AUI/A8345234/descendants'path Retrieve descendants of a UMLS atom and returns a JSON Object classType of Atom '/content/current/AUI/A8345234/parents'path Retrieve parents of a UMLS atom and returns a JSON Object classType of Atom '/content/current/AUI/A8345234/children'path Retrieve children of a UMLS atom and returns a JSON Object classType of Atom

'/content/current/CUI/C0155502/definitions'path Retrieves definitions of the CUI and returns a JSON Object classType of Definition

'/content/current/CUI/C0009044/relations' path Retrieves NLM-asserted relationships of the CUI and returns a JSON Object classType of ConceptRelation

'/content/current/source/SNOMEDCT_US/9468002'path Retrieves Source Concept and returns a JSON Object classType of SourceAtomCluster '/content/current/source/MSH/D015242'path Retrieves Source Descriptor and returns a JSON Object classType of SourceAtomCluster '/content/current/source/LNC/54112-8'path Retrieves Code and returns a JSON Object classType of SourceAtomCluster '/content/current/source/SNOMEDCT_US/9468002/atoms'path Retrieve atoms in a source-asserted identifier and returns a JSON Object classType of Atom '/content/current/source/SNOMEDCT_US/9468002/parents'path Retrieve immediate parents of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster '/content/current/source/SNOMEDCT_US/9468002/children'path Retrieve immediate children of source-asserted identifier and returns a JSON Object classType of SourceAtomCluster '/content/current/source/SNOMEDCT_US/9468002/ancestors'path Retrieve all ancestors of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster '/content/current/source/SNOMEDCT_US/9468002/descendants'path Retrieve all descendants of source-asserted identifier and returns a JSON Object classType of SourceAtomCluster '/content/current/source/SNOMEDCT_US/9468002/attributes'path Retrieves information about source-asserted attributes and returns a JSON Object classType of Attribute

'/content/current/source/SNOMEDCT_US/9468002/parents'path Retrieves parents a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster* '/content/current/source/SNOMEDCT_US/9468002/children'path Retrieves children of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster*

'/content/current/source/SNOMEDCT_US/9468002/parents'path Retrieves parents a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster* '/content/current/source/SNOMEDCT_US/9468002/children'path Retrieves children of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster*

'/content/current/source/SNOMEDCT_US/9468002/ancestors'path Retrieves ancestors of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster* '/content/current/source/SNOMEDCT_US/9468002/descendants'path Retrieves descendants of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster*

'/content/current/source/SNOMEDCT_US/9468002/ancestors'path Retrieves ancestors of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster* '/content/current/source/SNOMEDCT_US/9468002/descendants'path Retrieves descendants of a source-asserted identifier and returns a JSON Object classType of SourceAtomCluster*

'/content/current/source/LNC/44255-8/relations'path Retrieves relationships of LOINC code 44255-8 and returns a JSON Object classType of AtomClusterRelation

'/subsets/current'path Retrieves information about all subsets from the current release and returns a JSON Object classType of Subset '/subsets/current/source/SNOMEDCT_US/6011000124106'path Retrieves information for a SNOMED CT subset and returns a JSON Object classType of Subset '/subsets/current/source/SNOMEDCT_US/6011000124106/members'path Retrieves members of a SNOMED CT subset and returns a JSON Object classType of SourceConceptSubsetMember '/subsets/current/source/SNOMEDCT_US/6011000124106/member/89361000119103' path Retrieves an individual member of a SNOMED CT subset and returns a JSON Object classType of SourceConceptSubsetMember

'/subsets/current'path Retrieves information about all subsets from the current release and returns a JSON Object classType of Subset '/subsets/current/source/SNOMEDCT_US/6011000124106'path Retrieves information for a SNOMED CT subset and returns a JSON Object classType of Subset '/subsets/current/source/SNOMEDCT_US/6011000124106/members'path Retrieves members of a SNOMED CT subset and returns a JSON Object classType of SourceConceptSubsetMember '/subsets/current/source/SNOMEDCT_US/6011000124106/member/89361000119103' path Retrieves an individual member of a SNOMED CT subset and returns a JSON Object classType of SourceConceptSubsetMember

'/content/current/source/SNOMEDCT_US/9468002/attributes'path Retrieves attributes of the SNOMED CT concept and returns a JSON Object classType of Attribute

/semantic-network/current/TUI/T109' path Retrieves TUI and returns a JSON Object classType of SemanticType

'/content-views/current' path retrieves information about all subsets from the current release and returns a JSON Object classType of ContentView '/content-views/current/CUI/C2711988' path Retrieves information for the SNOMED CT CORE Problem List content view and returns a JSON Object classType of ContentView '/content-views/current/CUI/C2711988/members' path Retrieves members of the SNOMED CT CORE Problem List content view and returns a JSON Object classType of SourceConceptContentViewMember '/content-views/current/CUI/C2711988/member/238788004' path Retrieves an individual member of the SNOMED CT CORE Problem List content view and returns a JSON Object classType of SourceConceptContentViewMember

'/content-views/current' path retrieves information about all subsets from the current release and returns a JSON Object classType of ContentView '/content-views/current/CUI/C2711988' path Retrieves information for the SNOMED CT CORE Problem List content view and returns a JSON Object classType of ContentView '/content-views/current/CUI/C2711988/members' path Retrieves members of the SNOMED CT CORE Problem List content view and returns a JSON Object classType of SourceConceptContentViewMember '/content-views/current/CUI/C2711988/member/238788004' path Retrieves an individual member of the SNOMED CT CORE Problem List content view and returns a JSON Object classType of SourceConceptContentViewMember

'/crosswalk/current/source/HPO/HP:0001947'path Retrieves all codes that share a UMLS CUI with HP:0001947 and returns a JSON Object classType of SourceAtomCluster '/crosswalk/current/source/HPO/HP:0001947?targetSource=SNOMEDCT_US'path Retrieves all SNOMEDCT_US codes that share a UMLS CUI with HP:0001947 and returns a JSON Object classType of SourceAtomCluster

See also